Fri, Aug 2, 2024, 7:40 AM +02:00

+++ coke/coke.py
@@ -1,9 +1,9 @@
-amount_due = 50.
+amount_due = 50
 print("Amount Due :", amount_due)
 
-while amount_due != 0. :
-    prompt = float(input("Insert Coin: "))
-    if prompt == 25. or prompt ==10. or prompt == 5.:
+while amount_due != 0 :
+    prompt = int(input("Insert Coin: "))
+    if prompt == 25 or prompt ==10 or prompt == 5 :
             amount_due = amount_due  - prompt
             print("Amount Due :", amount_due)
     else:
